MAINT: stats: move multiscale_graphcorr
