+++ scipy/stats/__init__.py
@@ -633,6 +633,8 @@ from ._fit import fit, goodness_of_fit
 from ._covariance import Covariance
 from ._sensitivity_analysis import *
 from ._survival import *
+from ._mgc import multiscale_graphcorr
+
 
 # Deprecated namespaces, to be removed in v2.0.0
 from . import (
