+++ scipy/stats/stats.py
@@ -37,5 +37,5 @@ def __dir__():
 
 def __getattr__(name):
     return _sub_module_deprecation(sub_package="stats", module="stats",
-                                   private_modules=["_stats_py"], all=__all__,
+                                   private_modules=["_stats_py", "_mgc"], all=__all__,
                                    attribute=name)
